@@ -1,44 +1,38 @@
-[tool.poetry]
+[project]
 name = "magic512bot"
 version = "0.1.0"
 description = ""
-authors = ["bassmatt <bassmatt0515@gmail.com>"]
+authors = [
+    {name = "bassmatt", email = "bassmatt0515@gmail.com"}
+]
 readme = "README.md"
-packages = [{include = "magic512bot"}]
-
-[tool.poetry.dependencies]
-python = "^3.11"
-discord-py = "^2.4.0"
-python-dotenv = "^1.0.1"
-sqlalchemy = "^2.0.32"
-pytest = "^8.3.2"
-requests = "^2.32.3"
-table2ascii = "^1.1.3"
-psycopg2-binary = "^2.9.10"
-sqlalchemy-utils = "^0.41.2"
-
-[tool.poetry.group.dev.dependencies]
-flake8 = "^7.1.1"
-black = "^25.1.0"
-autopep8 = "^2.3.2"
-isort = "^6.0.0"
+requires-python = ">=3.11"
+dependencies = [
+    "discord-py>=2.4.0",
+    "python-dotenv>=1.0.1",
+    "sqlalchemy>=2.0.32",
+    "pytest>=8.3.2",
+    "requests>=2.32.3",
+    "table2ascii>=1.1.3",
+    "psycopg2-binary>=2.9.10",
+    "sqlalchemy-utils>=0.41.2",
+    "ruff>=0.3.3",
+    "mypy>=1.15.0",
+]
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.autopep8]
-max_line_length = 88
-aggressive = 3
-in-place = true
-recursive = true
-
-[tool.black]
+[tool.ruff]
 line-length = 88
-target-version = ['py38']
-include = '\.pyi?$'
+target-version = "py311"
+fix = true
 
-[tool.isort]
-profile = "black"
-multi_line_output = 3
-line_length = 88
+[tool.ruff.lint]
+select = ["E", "F", "I", "N", "W", "B", "UP", "PL", "RUF"]
+ignore = ["PLR0913"]  # Ignore too many arguments for now
+
+[tool.ruff.lint.isort]
+known-first-party = ["magic512bot"]
+combine-as-imports = true
